Total modification of execise functionality
Since it is stored in the txt file in the form of '<Exercise Name> - <Calculum Burns per Minute> kcal', it has been modified to import the contents with caution.
When the user enters the exercise number or time, the exception processing is set so that only the correct number may be entered, and a warning is output accordingly.