Correction for compilation error resolution
A compilation error occurred because the name of the structure 'member' used in the header file and the name of the member of execise did not match. This is resolved by member name matching.

Error using pointer due to type mismatch problem. execise_count is incorrectly used like a string in an integer variable or code. To solve this problem, use execise_count exactly as an array index and store new exercise information in the execises array. Modified to use the correct pointer when using strcpy