General modification of healthdata.c
The storage function for health_data is declared (saveData), but it has not been implemented. Since the saveData function is not actually implemented to store the data of the health_data structure in a file, there has been a problem that records are not stored in the health_data.txt file. The for statement is actively used to allow each exercise and diet input to be stored and imported into the txt file